Implement "like" feature for posts
Introduced the 'isLiked' attribute in the Post model to keep track of user's post liking. Updated the relevant GraphQL schema and adapted associated functions to manage reactions on posts. Adjusted the PostViewModel in order to properly handle post liking and unliking by updating the 'isLiked' status and the count of likes. User interface now displays feedback when a post is liked by the user.